+++ src/registry/proxy/remote_image/mod.rs
@@ -1,9 +1,12 @@
+mod download;
+
 use std::borrow::Cow;
 use std::fmt;
 
 use const_format::formatcp;
 use lazy_static::lazy_static;
 
+use crate::registry::SingleRegistryProxyConfig;
 use crate::registry::digest::{Digest, DigestError};
 use crate::registry::manifest::ManifestReference;
 
@@ -36,7 +39,8 @@ const fn get_image_ref_regex() -> &'static str {
 }
 
 #[derive(Clone, Debug, PartialEq)]
-pub struct RemoteImage {
+pub struct RemoteImage<'a> {
+    proxy_config: Option<&'a SingleRegistryProxyConfig>,
     /// `http` or `https`
     pub scheme: &'static str,
     /// Including port, docker.io by default
@@ -46,18 +50,19 @@ pub struct RemoteImage {
     pub reference: ManifestReference,
 }
 
-impl std::default::Default for RemoteImage {
+impl<'a> std::default::Default for RemoteImage<'a> {
     fn default() -> Self {
-        RemoteImage {
+        Self {
             scheme: "https",
-            host: "registry-1.docker.io".to_string(),
+            host: "docker.io".to_string(),
             repo: "(none)".to_string(),
             reference: ManifestReference::Tag("latest".to_string()),
+            proxy_config: None,
         }
     }
 }
 
-impl From<RemoteImage> for oci_client::Reference {
+impl<'a> From<RemoteImage<'a>> for oci_client::Reference {
     fn from(val: RemoteImage) -> Self {
         match val.reference {
             ManifestReference::Digest(d) => {
@@ -68,37 +73,35 @@ impl From<RemoteImage> for oci_client::Reference {
     }
 }
 
-impl fmt::Display for RemoteImage {
+impl<'a> fmt::Display for RemoteImage<'a> {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         write!(f, "{}", self.get_ref())
     }
 }
 
-impl RemoteImage {
-    pub fn new(mut host: &str, mut repo: String, reference: ManifestReference) -> Self {
-        if host.ends_with("docker.io") {
-            // The real docker registry is `registry-1.docker.io`, not `docker.io`.
-            host = "registry-1.docker.io";
-            if !repo.contains('/') {
-                // handle images like "nginx:latest" that are actually library/nginx:latest
-                repo = format!("library/{}", repo)
-            }
+impl<'a> RemoteImage<'a> {
+    pub fn new(
+        host: String,
+        mut repo: String,
+        reference: ManifestReference,
+        proxy_config: Option<&'a SingleRegistryProxyConfig>,
+    ) -> Self {
+        if host.ends_with("docker.io") && !repo.contains('/') {
+            // handle images like "nginx:latest" that are actually library/nginx:latest
+            repo = format!("library/{repo}")
         }
-
-        // Handle http:// and https:// in the repo uri
-        let scheme = if host.starts_with("http://") {
-            host = host.trim_start_matches("http://");
-            "http"
-        } else {
-            host = host.trim_start_matches("https://");
-            "https"
-        };
+        let insecure = matches!(
+            proxy_config,
+            Some(SingleRegistryProxyConfig { insecure: true, .. })
+        );
+        let scheme = if insecure { "http" } else { "https" };
 
         RemoteImage {
-            host: host.to_string(),
+            host,
             repo,
             reference,
             scheme,
+            proxy_config,
         }
     }
 
@@ -106,7 +109,7 @@ impl RemoteImage {
         &self.host
     }
 
-    /// Example return value: `https://registry-1.docker.io/v2/library/nginx`
+    /// Example return value: `https://docker.io/v2/library/nginx`
     pub fn get_base_uri(&self) -> String {
         format!("{}://{}/v2/{}", self.scheme, self.host, self.repo)
     }
@@ -115,7 +118,7 @@ impl RemoteImage {
         format!("{}/manifests/{}", self.get_base_uri(), self.reference)
     }
 
-    /// Example return value: `registry-1.docker.io/library/nginx@sha256:12345`
+    /// Example return value: `docker.io/library/nginx@sha256:12345`
     pub fn get_ref(&self) -> String {
         let (sep, ref_) = match &self.reference {
             ManifestReference::Digest(d) => ("@", Cow::Owned(d.to_string())),
@@ -128,8 +131,11 @@ impl RemoteImage {
         &self.repo
     }
 
+    pub fn get_proxy_cfg(&self) -> &Option<&'a SingleRegistryProxyConfig> {
+        &self.proxy_config
+    }
+
     /// Note: this uses the same validation rules as the Docker engine.
-    /// Schemes (`http`, `https`) are not supported here.
     pub fn try_from_str(image_ref: &str) -> Result<Self, RemoteImageError> {
         lazy_static! {
             static ref RE: regex::Regex = regex::RegexBuilder::new(get_image_ref_regex())
@@ -153,7 +159,7 @@ impl RemoteImage {
                 && &name[..i] != "localhost"
                 && name[..i].to_lowercase() == name[..i])
         {
-            host = "registry-1.docker.io";
+            host = "docker.io";
             repo = name.to_string();
         } else {
             host = &name[..i];
@@ -169,7 +175,7 @@ impl RemoteImage {
             ManifestReference::Tag("latest".to_string())
         };
 
-        Ok(Self::new(host, repo, ref_))
+        Ok(Self::new(host.to_string(), repo, ref_, None))
     }
 }
 
@@ -183,7 +189,7 @@ mod test {
         assert_eq!(
             ret,
             RemoteImage {
-                host: "registry-1.docker.io".to_string(),
+                host: "docker.io".to_string(),
                 repo: "library/debian".to_string(),
                 ..Default::default()
             }
@@ -192,7 +198,7 @@ mod test {
         assert_eq!(
             ret,
             RemoteImage {
-                host: "registry-1.docker.io".to_string(),
+                host: "docker.io".to_string(),
                 repo: "amouat/network-utils".to_string(),
                 ..Default::default()
             }
@@ -201,17 +207,17 @@ mod test {
         assert_eq!(
             ret,
             RemoteImage {
-                host: "registry-1.docker.io".to_string(),
+                host: "docker.io".to_string(),
                 repo: "amouat/network-utils".to_string(),
                 reference: ManifestReference::Tag("beta".to_string()),
                 ..Default::default()
             }
         );
-        ret = RemoteImage::try_from_str("registry-1.docker.io/mandy").unwrap();
+        ret = RemoteImage::try_from_str("docker.io/mandy").unwrap();
         assert_eq!(
             ret,
             RemoteImage {
-                host: "registry-1.docker.io".to_string(),
+                host: "docker.io".to_string(),
                 repo: "library/mandy".to_string(),
                 ..Default::default()
             }
@@ -286,7 +292,7 @@ mod test {
         for i in invalid_images.iter() {
             let ret = RemoteImage::try_from_str("http://docker.io/amouat/myimage:test");
             if let Ok(img) = ret {
-                panic!("Invalid image ref `{}` parsed as `{}`", i, img);
+                panic!("Invalid image ref `{i}` parsed as `{img}`");
             }
         }
     }
@@ -294,23 +300,26 @@ mod test {
     #[test]
     fn test_get_uri() {
         let img = RemoteImage::new(
-            "registry-1.docker.io",
+            "docker.io".to_string(),
             "debian".to_string(),
             ManifestReference::try_from_str("funky").unwrap(),
+            None,
         );
-        assert_eq!(
-            img.get_base_uri(),
-            "https://registry-1.docker.io/v2/library/debian"
-        );
+        assert_eq!(img.get_base_uri(), "https://docker.io/v2/library/debian");
         assert_eq!(
             img.get_manifest_url(),
-            "https://registry-1.docker.io/v2/library/debian/manifests/funky"
+            "https://docker.io/v2/library/debian/manifests/funky"
         );
 
+        let proxy_cfg = SingleRegistryProxyConfig {
+            insecure: true,
+            ..Default::default()
+        };
         let img = RemoteImage::new(
-            "http://cia.gov",
+            "cia.gov".to_string(),
             "not-watching".to_string(),
             ManifestReference::try_from_str("i-swear").unwrap(),
+            Some(&proxy_cfg),
         );
         assert_eq!(img.get_base_uri(), "http://cia.gov/v2/not-watching");
         assert_eq!(
@@ -319,13 +328,10 @@ mod test {
         );
 
         let img = RemoteImage::try_from_str("spy:v3.1.0-cia-INTERNAL").unwrap();
-        assert_eq!(
-            img.get_base_uri(),
-            "https://registry-1.docker.io/v2/library/spy"
-        );
+        assert_eq!(img.get_base_uri(), "https://docker.io/v2/library/spy");
         assert_eq!(
             img.get_manifest_url(),
-            "https://registry-1.docker.io/v2/library/spy/manifests/v3.1.0-cia-INTERNAL"
+            "https://docker.io/v2/library/spy/manifests/v3.1.0-cia-INTERNAL"
         );
     }
 }