+++ src/routes/extracts.rs
@@ -5,8 +5,14 @@ use axum::extract::{FromRef, FromRequestParts};
 use axum::http::StatusCode;
 use axum::http::request::Parts;
 use axum_extra::extract::Host;
+use serde::Deserialize;
 
 use crate::TrowServerState;
+
+#[derive(Deserialize)]
+pub struct ImageNamespace {
+    pub ns: Option<String>,
+}
 
 pub struct AlwaysHost(pub String);
 
@@ -31,7 +37,7 @@ where
                 "http"
             };
 
-            return Ok(AlwaysHost(format!("{}://{}", scheme, host)));
+            return Ok(AlwaysHost(format!("{scheme}://{host}")));
         }
 
         Ok(AlwaysHost(String::new()))