+++ tests/authentication.rs
@@ -67,7 +67,7 @@ mod authentication_tests {
             .clone()
             .oneshot(
                 Request::get("/login")
-                    .header(header::AUTHORIZATION, format!("Basic {}", bytes))
+                    .header(header::AUTHORIZATION, format!("Basic {bytes}"))
                     .body(Body::empty())
                     .unwrap(),
             )